+++ myems-api/core/energystoragepowerstation.py
@@ -860,20 +860,6 @@ class EnergyStoragePowerStationImport:
     def on_post(req, resp):
         """Handles POST requests"""
         admin_control(req)
-        # try:
-        #     upload = req.get_param('file')
-        #     # Read upload file as binary
-        #     raw_blob = upload.file.read()
-        # except Exception as ex:
-        #     raise falcon.HTTPError(status=falcon.HTTP_400, title='API.ERROR',
-        #                            description='API.FAILED_TO_UPLOAD_IMPORT_FILE')
-        #
-        # try:
-        #     raw_json = raw_blob.decode('utf-8')
-        # except Exception as ex:
-        #     raise falcon.HTTPError(status=falcon.HTTP_400,
-        #                            title='API.BAD_REQUEST',
-        #                            description='API.FAILED_TO_READ_REQUEST_STREAM')
         try:
             raw_json = req.stream.read().decode('utf-8')
         except Exception as ex: